Release v0.2.0 - Production-Ready Pure-Go Race Detector

Major achievements:
- 99% overhead reduction (15-22% vs typical 5-15x)
- 670+ tests passing with 99%+ coverage
- Complete race detection with full stack traces
- Zero external dependencies (Pure Go)
- Cross-platform support (Linux, macOS, Windows)
- CI/CD integration ready

See CHANGELOG.md for full details.

+++ internal/race/detector/detector_bench_test.go
@@ -98,7 +98,7 @@ func BenchmarkOnWrite_WithRace(b *testing.B) {
 		// Set previous write to future epoch to trigger race.
 		vs.W = epoch.NewEpoch(1, 1000000)
 		ctx.C.Set(1, uint32(i))
-		ctx.Epoch = epoch.NewEpoch(1, uint32(i))
+		ctx.Epoch = epoch.NewEpoch(1, uint64(i))
 
 		d.OnWrite(addr, ctx)
 	}
@@ -345,7 +345,7 @@ func BenchmarkOnRead_WithRace(b *testing.B) {
 		// Set previous write to future epoch to trigger race.
 		vs.W = epoch.NewEpoch(1, 1000000)
 		ctx.C.Set(1, uint32(i))
-		ctx.Epoch = epoch.NewEpoch(1, uint32(i))
+		ctx.Epoch = epoch.NewEpoch(1, uint64(i))
 
 		d.OnRead(addr, ctx)
 	}
